Moved IR type testing back into IrObserver implementation classes, out of IrObserver interface.
Added support for OptionSet instances to print a header in the help message before the help for each option.

Modified all programs that use the PrototypeGenerator so that they include the PrototypeGenerators options in their own. This means that the help message now describes *all* the options available for the program.