Normalized the identification of threads in the inspector. Threads can now be identified by the VM issued ID (i.e. their ID in the VMThreadMap) or the handle used by the native thread library (e.g. LWP ID in Solaris).
Moved the up call when gathering threads to TeleProcess.jniGatherThread to centralize handling threads in a semi-initialized state. That is, threads for which the ID is known but not yet *all* the VM thread local spaces. This can be seen by the inspector if a thread is stopped before executing the relevant native code that initializes these members in the associated ThreadSpecifics struct.