Extended compilation spec accepted by C1XTest to use patterns. See output of 'max help c1x'.
Added @Override annotation to methods in C1X and enabled checking of this annotation in the C1X Eclipse project.

Completed printing of Constant instructions by InstructionPrinter.

Extended meaning of 'prototype only' classes (as implemented by MaxineVM.isPrototypeOnly()) to exclude Maxine classes that are in a MaxPackage which is not a BasePackage, AsmPackage, VMPackage or test package.

Added -o flag to C1XTest so that all output can be sent to a file. This is particularly useful when running inside Eclipse.

Added -print-ir flag to C1XTest so turn on IR printing of all (successfully) compiled methods.