Modified JNI function epilogue to re-use JNI call prologue as the semantics with respect to GC/mutator synchronization are identical.
Removed notion of "hard" safepoint as a) it is no longer used and b) its intended semantics cannot be implemented.

Improved mutex and conditon variable tracing in C code.

Modified JitTargetMethod.finalizeReferenceMaps() so that it now uses a spin-loop instead of synchronization to ensure only one thread does the reference map finalization. The problem with synchronization is that is can block a thread on a mutex in native code which the GC will falsely interpret as meaning the thread has safely been stopped for the purpose of not mutating during a GC,

Fixed bug in verifier when handling version 50.0 classfiles that fail type-checking verification but pass when failing-over to type-inferencing verification.