Modified BlockingServerDaemon to separate initialization from starting of the server thread.
Added Log.printVmThreadLocals() to support logging of VM thread locals.

Added documentation for more SpecialBuiltins.

Moved creation of stick monitors closer to the declaration of the boot image objects to which they are bound.

Added more documentation to the monitor code.

Added support (in Throw.java) for dumping the stack of (stopped) thread from a different thread.

Fixed bug in trap stub code where a safepoint trap occurred on a thread that has a null safepoint procedure. In this case, the safepoint latch register was not being reset to the safepoints-enabled thread local space; now it is.

Changed tracing in Object_new benchmark so that only one thread traces its progress.