Moved a number of static fields of type ClassActor, FieldActor or MethodActor into ClassRegistry. This is convenient location for these well known symbols and also makes it easier to control the order in which they are initialized to class initialization at bootstrap time.
Fixed a bug in CPS backend to do with the interaction between register allocation and stack variable pinning (exercised by the MakeStackVariable builtin).

More changes towards making the AttachCurrentThread JNI function work.

Made the default external viewer in the Inspector be SOCKET with port 2341 so that EclipseCall works out of the box.