Modified protocol when starting a new thread such that the starter is no longer blocked while the startee is running.
Removed RunScheme.run() method offset from boot image.

Expanded thread state in the Inspector to include both local thread handle info and remote thread handle info. The former is the handle in the Inspector's address space provided by the platform-specific process control API. The latter is the opaque handle to the native thread in the remote VM (e.g. the value of pthread_self()).

Disabled the native method dialog by default when stopping or switching to a thread whose current code location is in a native function. You can still bring up the dialog by using the popup menu of a frame in the stack inspector.

Removed all /* non-Javadoc ... */ comments as they are redundant in all modern IDEs that support navigating from a method to the interface method it implements.

Added new VM_ENTRY_POINT annotation so that C_FUNCTION is no longer confusingly overloaded. This has also replaced the JNI_FUNCTION annotation.