Implemented shared adapters. The sharing is signature based. Modifications for SPARC adapters not yet complete.
Added infrastructure so that Inspector can iterate over all the target methods in the VM's code manager. A new dialog was added using this that allows the user to select from a list of all target methods. It's availble via the new 'Code -> View target mode...' menu item (hotkey: CTRL-D). This dialog enables one to view target code that is not associated with a class method actor (e.g. shared adapters).