Improved code emitted for pointer loads and stores by doing peephole constant folding of complex addresses.
Removed null checks when dereferencing pointer returned by ALLOCSTKVAR and ALLOCA instructions.

Stopped backend from emitting loads of constants into registers that are never subsequently used.

Fixed Inspector bug when obtaining a name for a stub which is a TargetMethod with a null ClassMethodActor.

Added 'ciKind' field to Kind to cache the CiKind for a Kind.