Added support for integrating with JDK 1.6.0_20 on Mac OS X (most recent version). This comes at the cost of obsoleting support for the previous version (JDK 1.6.0_17) as supporting both introduces more complexity than is justifiable.
Enhanced C1Visualizer output to include reference maps in LIR traces.

Fixed bug in C1X AMD64 global stub emitter where the register state area being saved to was not aligned with the register state area scanned by the GC. This only showed up on Darwin due to 16-byte stack frame alignment issues.

Fixed bug in C1X XIR stubs for runtime calls to methods returning boolean, short, byte or char. In all these case, a full 32-bit value needs to be copied to the stack slot used to return the value from the stub to the XIR snippet.

Added support to Log.java and log.[ch] for printing the symbol information (if nay) for an address vi dladdr(3). This is now used in crash dumps to show any native functions on a thread's stack.

Improved reporting of unimplemented functions in jvm.c to include a stack trace (if possible).

Downgraded tracing of method substitution to trace level 2.