Adopted HotSpot mechanism for executing VM operations on a dedicated thread. This involved some recasting and generalization of FreezeThreads and StopTheWorldGCDaemon to become VmOperation and VmOperationThread respectively.
Implemented support for user provided Java signal handlers. Used this support to implement thread dumping with SIGQUIT is sent to the VM.

Fixed linker issue on Darwin where the JDK libraries (as of 1.6.0_20) are linked against a library named libjvmlinkage.dylib. This is now the name of the Maxine library generated for the native Maxine code (on Darwin only).

Made the global thread lock a separate object (VmThreadMap.THREAD_LOCK) separate from the global thread map (VmThreadMap.ACTIVE). This makes it clearer to see where this lock is taken.

Made VM options be sorted primarily by category (standard, non-standard, implementation-specific or C1X-specific) and secondarily by name. This makes it easier to search for the options in the VM's various usage messages.