Added pre-built VmThread and java.lang.Thread objects to the boot image for special VM threads that may be referenced during GC.
Converted a number of uses of VMBooleanXXOption/VMIntOption/VMStringOption/VMFloatOption into uses of VMOptions.addFieldOption(...) instead. Also changed the latter to be based on java.lang.reflect.Field instead of FieldActor so that this facility can be used in classes whose initializers are run before ClassActors (and thus FieldActors) can be created.

Used-defined, Java-based signal handling (i.e. sun.misc.Signal and friends) now works. This is used by the VM to install a SIGHUP handler that triggers a stack dump of all the threads in the system. The PrintThreadsDemo class demonstrates a way to modify this behavior.

Modified the stack walker so that a more complete stack is shown in the Inspector for the VM operation thread when it receives an asynchronous signal while it was executing/blocked in native code.

Introduced VmOperation.submit() method as convenient short-hand for VmOperationThread.submit(vmOperation). It also gives VmOperation subclasses an extension point.

Turned off AsynchronousProfiler.