Fixed MS heap scheme regression due recent removal of contiguous space requirement and use of non-fixed code manager.
MS heap does requires the heap to be above the boot and code region (although, not necessarily contiguous).
Making them contiguous such that boot < code < heap allows to easily fulfill this requirement.
This ordering is required to make marking efficient by reducing the amount of testing needed
at each visited reference to determine if a the reference is covered by the mark-bitmap.

This requirement for contiguous heap is implemented by storing in the boot image a bootImageReservedSpace that
specify whether the boot region should be map at the beginning of a reserved range of virtual address of size
specified by bootImageReservedSpace.
At prototyping time, the value of bootImageReservedSpace is queried from the HeapScheme (thus this decision is HeapScheme
specific).
At boot image load time, if the value of bootImageReservedSpace isn't 0, a mapping is created (on Unix, with MAP_NO_RESERVE).
Use/disposal of part or whole of the mapping is the responsibility of the Heap Scheme.

Also included here, the reversal to FixedAddressCodeManager for Solaris / Darwin.