Separated OS level signal handling path for VM signals (i.e. those handled directly by the VM for implementing safepoints, runtime exceptions, de-opt etc) and Java signals (i.e. those dispatched to registered instances of sun.misc.SignalHandler).
Added -XX:+SerializeSignals option to use a single thread for executing each
sun.misc.SignalHandler (instead of starting a new thread for each signal). This is
a work around for the fact that there is a scalability issue with thread creation
in Maxine. Creating thousands of threads in quick succession (e.g. upon receiving
thousands of SIGHUPs) results in increasing slower thread creation.