Moved definition of register reference map bit count from CiRegisterSaveArea to CiArchitecture as it is fixed for an architecture, not per callee save area layout. The latter interprets the bit map in terms of exactly where it stores the denoted registers.
Made CiRegisterSaveArea be an attribute of an RiRegisterConfig, not a CiArchitecture.

Extend LIRAddressOperand to be able to model a LIR address where the base register is fixed, but the index register is not (prior to register allocation).

Re-enabled encoding of de-opt frame info into CiDebugInfo objects in C1X backend.