Fixed the enforcement of the requirements for patching of direct calls and jumps in CPS and C1X, on amd64. A direct call in maxine for amd64 is always a call to a 4-byte displacement. A direct call instruction thus always takes 5 bytes. When patching a direct call, we only need to change the displacement part of the instruction (i.e., the trailing 4 bytes).
There are two requirements:
- do not straddle a cache line. On x64, a cache line is most likely to be some power of 2 times 16 bytes
  (i.e., a double-word * 2^p).
- update the displacement in the call instruction with a single memory write.

The second requirement means that the displacement should be 4-bytes aligned.

Both requirements can be simply achieved if direct calls fits in an 8 byte-aligned chunk
and is laid out such that the call instruction is at the 3 bytes (starting from 0).
This means that at worst a direct call instruction might be preceded by up to 7 nops.

What complicates the picture for CPS is that the assembler allows inclusion of variable length instructions whose final length
is decided when the method is completely generated. Any instruction with alignment requirement add an Alignment directive
that reserve enough space to fill it with nop to preserve the alignment. The interface makes it difficult to specify that
we need the call instructions to 3 bytes away from alignment, and force us systematically prefix the call instruction with
3 nops in addition to those added by the directive to enforce the requirements. This means we can end up with up to 10 nops !!!

We'll fix this later by:
1. supporting patching of unaligned displacement;
2. adding a "containment" directive to CPS's that allows an instruction to float anywhere inside a alignment constraints.

Additional changes:
- Renamed the utility class AMD64TargetMethod to the more appropriate AMD64TargetMethodUtil class and move this to the
cps-independent compiler.target.amd64 package.
- Reworked the patching interface of TargetMethod:
  1. to re-use code patching support across CPS / C1X
  2. to make explicit the distinction between (i) fixing up code immediately after compilation, or anytime before any thread
     can execute the code and (ii) mt-safe patching of code that may be concurrently executed by threads.
     There are much more occurrences of the former which doesn't requires the complexity of code patching. Some places
     in code are never subjected to runtime patching and therefore shouldn't be required any alignment constraints for patching
     purposes.

     For now, the only real case of runtime patchings I've come across in Maxine are: (i) static trampoline, (ii) method forwarding.
     Linking of direct calls (the vast majority of code updates) are fixup only (although the alignment need to be preserve as
     this may fixed up into static trampoline calls).