Made C1X the default -opt compiler when configuring a boot image.
Made it possible to built a boot image with C1X as the -opt compiler and the template JIT compiler as the -jit compiler. This is now the default boot image.

Made C1X translate the Safepoint builtin as a no-op.

Added support for field-based VM options of the form "-name" or "-Xname".