Added T1X, a template JIT whose templates are compiled by C1X.
Renamings:
    'bytecodePosition' -> 'bci'
    'targetCodePosition' -> 'pos'

Added GUI option to 'max methodtree'. To use it: 'max methodtree -view'. This was inspired by the GUI for 'max objecttree'.

Added support to C1X for emitting jump tables when translating tableswitch JVM instructions.

Split single compilation event trigger into two: one trigger for compilation start events and one for compilation end events. The former is only enabled when the -precomp-bp Inspector option is specified and so the trigger overhead is avoided when this option is not used.

Added support to CRI for passing code annotations back to the runtime. Code annotations give extra information about a position in the code of a CiTargetMethod. The annotations currently generated by C1X are code comments (when -C1X:+CommentedAssembly is used) and jump table descriptors. The latter are used by the Maxine disassembler to format jump tables nicely.

Removed CPS and C1X as dependencies of the Tele and Inspector projects.

Moved more CPS specific classes into the CPS project.