Replaced the compiled methods that unwind the stack when raising an exception with hand generated stubs.
Made finalizer registration only happen when returning from Object.<init>. Previously it was being done when allocating an object which is incorrect.

Made the default optimization level for C1X (in Maxine) *really* be 3. The previous attempt at this only set the default value for the -C1X:OptLevel flag which is not the same thing.

Improved the names of certain methods in RiRuntime.

Substituted some methods in java.lang.re.ReferenceQueue so that operations adding/removing elements to the queue can be traced with -XX:+TraceReferenceGC.

Enhanced ALIAS so that exact matching of field/return/parameter types is not required when aliasing (although type safety is still guaranteed).

Removed C1XOptions.GenFinalizerRegistration (this is now always "true").