Modified generation of stubs by C1X to use register configurations supplied by the runtime. This fixes a deopt bug where a stack slot being used to pass a stub parameter overlapped with a slot being used to save an original return address during deopt.
Renamed CiCalleeSaveArea to CiCalleeSaveLayout to better matched the role of the class. A "callee save area" is now uniformly a pointer value denoting the actual memory storing the callee saved registers.

Renamed GlobalStub to CompilerStub to be less ambiguous in the namespace of the various different stub types in a VM.

Simplified processing of objects references in registers during stack reference map preparation. This caused TargetMethod.Flavor to be removed and replaced with Stub.Type.