Refactored a number of projects to migrate to the use of the com.oracle.max package prefix:     com.sun.max.vm.compiler.c1x -> com.oracle.max.vm.ext.c1x     com.sun.max.vm.t1x -> com.oracle.max.vm.ext.t1x     com.sun.max.vm.bytecode.transform -> com.oracle.max.vm.ext.bctrans
Made sure all VmThreadLocal objects not created as part of scheme instantiation are referenced in the constructor of the BootImagePackage subclass responsible for loading them in the boot image. This ensures that the complete set of VmThreadLocals are known before any operations on the set (such as VmThreadLocal.tlaSize()) are performed.