Completed refactoring of refill managers to split them out of the first-fit mark-sweep space. This will sharing them across different allocators implementation used to parameterize the mark sweep space, allowing further re-use of the mark-sweep space between flat and generational heap schemes.
Move card-table implementation into a sub-package of the gcx package. This is the first step in the re-organization of the gcx package into multiple smaller packages in order to reduce dependencies between the various components and to avoid sucking in all gcx classes into every heap scheme using a subset of it.