Add aarch64 class file and its register structure.
This commit is only an initial experimental one because exceptions and error will occur.

Register structure of aarch64 is quite different from that of arm64. The following difficulties were met.
1. Register flags like "Byte" are needed to pass the VM configuration generation. But registers in aarch64 can only be used in full 64-bit or lower 32-bit.
2. I am not yet very clear about aarch64 registers' meaning and usage, for example, frame pointer register and LATCH register are needed by maxine, but I do not know how to configure them correctly.

I was trying to first ingore all the detailed implementation but sometimes the maxine just needs an instance as a return value for example. In this case I would have to give it a patchy implementation. However it turned out to be a very large amount of work...