This "fixes" the crash due to java.util.Random.seedUniquifier The compareAndSet resulting in a compareAndSwapLong faults as mulLong in ARMV7Assembler had a push without a pop, this corrupted stack.
Also fixed, compareAndSwapLong (untested) which used compareAndSwapWord in
JDK_sun_misc_Unsafe.java,  this is fine for X86 on 64 bits, but is a problem for
ARM32Bit. Added the additional "Reference" functions based on the same ideas as
compareAndSwapInt

We need to test compareAndSwapLong.

Current problem: we fail to laod the main "Hello" class at the moment. Unidentified cause of problem at the
moment, but could be because of compareAndSwapLong -- we need to check there is/are no retrievals of objects that
might be failing due to 32bit object id issues.

Apologies if these edits are not conformant to coding guidelines. They were done using vi.
A subsequent commit will fix this shortly.