Fix to I2C (trivial error) Fix to Stubs.genUnwind for ARMV7 (not sure stack address is constructed correctly) but seems to work
Current issues, we sometimes get a BUS error, due to an unaligned lrdexd, (needs to be 8 byte aligned),
tried AtomicLong replacement (not included) but this did not fix it. Rebuilding the image with a different number of threads sometimes fixes the problem.


Fails during load of InetAddress.java seems to be related to com.sun.max.vm.classfile.ClassfileReader.readInnerClassesAttribute
where we get a clas cast exception during construction of InnerClassInfo