Added support for DEBUG_METHOD for X86 see x86.py Unfortunately, we have a strange problem related to resolveVirtualCall if we remove the Log.printlns from this method then we get a SIGSEGV with a null caller address. I tried being more conservative with barriers in ARMV7.java --- this might need to be reverted.
The problem of the SIGBUS for the ldrdex is becoming more frequent during compilations.
We also get problems where compilation fails -- this seems to be a register allocator bug due to intervals? Using more compilation threads
makes it more likely for this bug to appear.


Apologies, I cannot get the commit any cleaner than this at the moment. It will need cleaning/formatting in an IDE.

I have not got a reasonable answer as to why inserting/removing the Log.println in  Stubs.java has such an effect.
I have done a fair bit of work on Stubs.java but mroe definitely needs to be done. Some handcrafted assembler is still wrong.