This includes --- I think the necessary fixes to allow ARM32Box to be sued to get around the hashCode problems for Compiler dependencies see validateDependenencies ....
I have also started to attempt to fix the patchStaticTrampolins fixup32CallSite in ARMTargetMEthodUtil for non-hosted mode,

What does this mean ---- the methods have been compiled and the trampolines need to be patched ....
So we are very close to debugging the Adapters ... the Stubs seems to "work" but are probably a bit buggy for exception handling
Notable issues include not pushing the flags during exception handling. Deoptimisation is likely to be wrong.

.... The CiAssumptions hashCode has BEEN PERMANENTLY hacked to ARM for 32 bitsi(wordSizeHack as I could not get access to Platform .. claimed package did not exist when I imported and compiled mx build)
 .... thw wordSizeHack needs to be removed by refactoring when the objects are created.
Also, the hashCode stuff might not be necessary ... I will try this ASAP and remove it if possible.