Fixes to ARMAdapters they now correctly output push(lr) movw movt add r12,r12,r15 blx r12 NOTE the changes to CallEntryPoint, ARMAdapterGenerator, and ARMV7LIRAssembler (emitIndirectCalli r8 offset was incremented by 4).
Fixes to genTrapStub ... now fixed completely I think. OTHER Stubs are still buggy.

Current issue --- getting a SIGBUS when we do
536871077 com.oracle.max.vm.ext.t1x.jvmti.JVMTI_T1XTemplateSource.resolveAndGetFieldDouble
This is a new one on me ... it occurs after patching of addresses so maybe we are getting further.
I will check to make sure I have not incorrectly patched something ....