Apologies for pushing this right now in case I have broken anything ... forever -- loops put into some parts of the code ...
IMPORTANT fix to branch that was causing a problem.
Issues -- Adapters are not called yet ... but some T1XTemplates from JVMTI are executed.
Adapter targets seem to be 8 bytes out (2 instructions past the beginning). Not sure if this is due to CallEntryPointer or if
it is due to another problem.