Sorry this is a bit dirty ... The adapters are being called, but the return addresses are getting stuffed up. Looking at the stack pointer. The return address on there points to 536873310 java.net.URLStreamHandler.hashCode The LR points to just past the BLX in the adapter of the BASEINE compiled 536870913 java.net.URLStreamHandler.getHostAddress
THe things that can go wrong are .. wrong entry points to the adapters themselves.
Incorrect stack manipulations.

Or work is slightly complicated by the fact that we must push LR on entry to a routine, and this is also our
entry point.

Consider the scenario we go from baseline2OPT, this means we enter at the adapter calling bit of code.
This pushes the LR onto the stack, and generates a call target address in r12 and blxs r12. The return from
the adapter will be the next instruction -- the main entry point to the method.
The adapters must add 4 to the return address and then do a ret, or an equivalent operation. Plus they must adapt the stack
appropriately, ie subtract and add. and push pop the frame pointer.

I will add some debugging to the adapters --- only when not hosted mode, to see what is happening. My only worry
is that why does C1X compilation of Hello world not work? Surely that will not require any adapters.