Code cleanup of debugging prints, a lot are temporarily put under verboseCompilation concerning patching trampolines, code fixup etc.
X86 now passes nearly all the tests, it might even be identical to what Andrey/kenai has.
I will now test this on ARM. I added one of our ARM bytecode tests to include HELLO WORLD ...
It is the first test run with integer identifier -1. We can try to put a more significant code example
in an appropriate place in the source code tree for the demo on Wednesday.