Fix to the  demo ... mx test -tests=javatester
Will create a bootimage with the SimpleExample compiled in, it only runs that test in the jtt-c1c1x tests.
It reads a file lena.raw, blurs it with a gaussian filter and produces a new lenablurred.raw file.
These files are prdouced in the maxine-tester/jtt-c1xc1x directory.
In this directory there is a python file, to view the images.

This revision should be the version we use for the demo.