536870914 java.net.URL.getHost One gets slightly different behaviour dependent on what arguments are used to execute under gdb I found using "run -verbose" useful.
The current problem is related to the adapters. The stack seems to be getting corrupted, in that under gdb with run -verbose the
return address on the stack when returning from the java.net.URL.getHost method is incorrect.

Static trampoline patching seems to function ok. We seem to enter at the correct address for a baseline/baseline method caller/callee.
The stack adaptation for OPT2BASE seemed to be ok.

I managed to reliably singly step and verify that the stack seemed to be correct upto and including slide 17 of the MaxineAdapters.pdf.
I verifed that r0 and r1 as input arguments to URLStreamHandler.getHostAddress(URL) matched up as the this-ptr and the URL-ptr and were placed
on the stack in expected places.


I did see code such as that below. Slightly concerned as the manipulation of the stack SP register in unusal places. THIS MIGHT BE A SOURCE OF ERRORS.

SUGGESTED STEPS:
1. Identify where the code sequence below came from. It looks like T1X code and is certainly being executed during a baseline compiled method. I don;t believe it
it the Adapter.

2. Check (possibly do this first), check the ARMV7T1X jtt tests and see if anything seems broken.

3. Single step. Identify (one entry to getHost) the LR and FP (R13 R11) values for the return address and framepointer and their location on the stack. Determine
where, when and why the stored values are either corrupted, and/or why the stack pointer value is incorrect.

  0x141d9334:	nop	{0}
   0x141d9338:	nop	{0}
=> 0x141d933c:	nop	{0}
   0x141d9340:	movw	r12, #52	; 0x34
   0x141d9344:	movt	r12, #0
   0x141d9348:	add	r12, r12, r11
   0x141d934c:	ldr	r8, [r12]
   0x141d9350:	movw	r12, #4
   0x141d9354:	movt	r12, #0
   0x141d9358:	sub	sp, sp, r12
   0x141d935c:	mov	r12, sp
   0x141d9360:	str	r8, [r12]
   0x141d9364:	mov	r12, sp
   0x141d9368:	ldr	r0, [r12]
   0x141d936c:	movw	r1, #44	; 0x2c
   0x141d9370:	movt	r1, #0
   0x141d9374:	add	r12, r0, r1
   0x141d9378:	ldr	r0, [r12, #-0]
   0x141d937c:	nop	{0}
   0x141d9380:	nop	{0}