Force execution of C1XOptions.GenSpecialDivChecks in ARMV7LIRAssembler to handle special case integer division as per the java specification. We can remove this, perhaps it is better to make it a command line option again.
The float replacementi function  for integer division is oki and passes signed integer division tests, , but it is running without vpush vpop at
the moment and that is a little bit dangerous.

I will fix vpush/vpop tomorrow and check to see if it allows us to build and execute on
the Zynq, there may be some issues with longs that I have not thought about, and/or
some hardcoded use of idiv.