Basic fixes to some encodings ... vcvt etc, Upgraded the float sdiv replacement to use doubles. mx vm now boots using the double sdiv/udiv replacement.
Issues: cannot build on the Zynq ... not enough memory ... John could add some swap but it will be pretty slow as it is only 600MHz and
it only has 1G memory.

So the plan for now is to build on ODROID and to add some settings to make it build for cortex-a8 or a9? And ensure it knows the integer divide is not there.
Hopefully this will be enough to then use the resulting binaries on the Zynq. If this works then we can then move onto the instrumentation of the LOAD/STORES.

We might need to enable SWAP just to get Maxine running.