Added some extra functionality into offlineT1XCompile. Most of the T1Xtests now pass. Current failures include Unexpected failures:   test_T1X_jtt_BC_XXXXfdiv(test.arm.jtt.ARMV7JTTTest) failed   test_T1X_jtt_BC_d2i02(test.arm.jtt.ARMV7JTTTest) failed EXPECTED USES AN ARRAY.   test_T1X_jtt_BC_getstatic_i(test.arm.jtt.ARMV7JTTTest) failed EXPECTED ACCESES static instance variable.   test_T1X_jtt_BC_ineg(test.arm.jtt.ARMV7JTTTest) failed   test_T1X_jtt_BC_ineg_1(test.arm.jtt.ARMV7JTTTest) failed   test_T1X_jtt_UsageOfStaticMethods(test.arm.jtt.ARMV7JTTTest) failed EXPECTED has System.out.println
NOTE: on the neg tests, it seems to bbe due to the second argument being uninitialised.
I seem to recall, in the past we manually set the zero argument to 0.
The failure is a java.lang error concerning uninitialised argument. I don't seem to understand how
this argument is initialised.

It might be possible to extend the T1X tests, to properly test the adapters. I guess it should be possible to enable the
generation of an OPT2Base Adapter and for return arguments to be passed back on the stack.