I'm doing a dirty push in case of a problem/issue with the odrodxu3a. Mandelbrot is hacked, we have a problem with float/double 2D and possibly 1D arrays.
Basic instrumentation support is in ....
It now needs adding to the appropriate ld/st instructions.
It merely prints to the screen for now.

I had to add an interface to get round the circular dependence.
I crudely call the C function to flush using a blx in the instrument method.
I will need to do some cleanup/updates as maxine.c erroneously reports an error on flushing at the moment.