nisbeta@odrodxu3a:/scratch/nisbeta/MAXINE/maxine$  mx vm   -cp  .:com.oracle.max.tests/bin Hello2 > XXX We appear to have a concurrency bug, sometimes this program will run to completion, other times it will crash. I appear to have enabled the various invoke and invoke virtuals to function correctly. My hunch is that maybe we have not correctly implemented instruction cache flushing (relating to T1X jit-compiled code), and/or there are other concurrency related issues.
This is still a fairly dirty push ie very mainy Log/prints included. I will try to examine if we have missed any operations that
relied on atomic operations/instructions in C1X for X86 that we have not appropriately replaced. I do not know if it is somehow
related to 20bit hashcodes and object retrieval but I do not think this is the case.