Current status, attempted some alignment fixes ... but the SIGBUS is generated by a method that is compiled at bootimage time where two long and one AtomicLong automatic variables are used to do a compare and swap ... ?
The current problem relates to the initialisation of the second argument register $r1 in T1X initated invokes, I need to determine the
correct place to saver and restore it .. or rewrite the code to perform initialisation of $r0 and $r1 to not overwrite it. The problem
is due to the fact that T1X has been harcoded for the integral return register to be different to the input argument registers.
1

java.util.Random ...
protected int next(int bits) {
        long oldseed, nextseed;
        AtomicLong seed = this.seed;
        do {
            oldseed = seed.get();
            nextseed = (oldseed * multiplier + addend) & mask;
        } while (!seed.compareAndSet(oldseed, nextseed));
        return (int)(nextseed >>> (48 - bits));
    }
This method is compiled at boot image generation time. It is an issue when running the Hello2.java
 mx vm   -verbose -XX:+TraceTraps   -cp  .:com.oracle.max.tests/bin Hello2
Adding/removing -verbose or TraceTraps is generally sufficient to cause a SIGBUS ....