OK ... so I rolled back/edited out the SIGBUS logging stuff just to see where we are upto. When I built the bootimage with the alignment logging in it crashed very early.
FIXES: vdiv ... incorrect encoding for double precision
	IDIV/UDIV/LDIV by zero exceptions should now work
	without false positives. We now example the contents of the PC and some "ILLEGAL INSTRUCTION"
	markers in the code prior to the vldr that is used to force an exception, it's then handled in
	Trap.java.

What is broken ... GC is causing most problems, it always SEGSEGVs.
Seems to be a problem with Thread interruption ...
Potentially an issue with barriers.
SIGBUS issues these are now a priority.

Other issues also affecting X86, we have introduced a bug since kenai, I tried to run a Mandelbrot and this works on
kenai but  fails on out x86. We need to investigate this using bisection of the revisions, we need to determine which revisions
will build bootimages for x86, and record them in order to aid subsequent bug fixing by revision bisection.

I will place the equivalent of an assert into the prologues of methods, adapters and  stubs to check that the stack frame after
allocation (subtraction) is 8byte aligned.