Fix to signal handler in substrate to account for Thumb bit set in CPSR Thread_isInterrupted02 fails like X86 and Thread_isInterrupted03 passes like X86
Getting some VMThread[Thread-25,5,main] prints that need to be removed from the VM.


Bernard>>>
Note that there is still a bug in this implementation, however it will
be extremely rare and is unlikely to be hit in any tests: if a thread
is interrupted in the middle of a Thumb IT (if-then) block then the
CPU will save the IT state machine in 8 bits of the CPSR. These bits
should be restored when control is returned to the interrupted code,
but the *only* way to set these bits on Linux is by using the
sigreturn system call (the CPSR bits can only be modified in
supervisor mode). Fixing this would require significant refactoring of
the way Maxine invokes trap handlers, so I don't think you should
worry about this just yet, but it is something to keep in mind.
<<<<<<<