[jdk7u] Add LD_LIBRARY_PATH to environment to make dlopen work
`DT_RPATH` is deprecated and replaced by `DT_RUNPATH`.  However
`DT_RUNPATH` is not transitive and thus it's not propagated through
maxvm to `dlopen` when loading `libjava.so`.  As a result, when loading
`libjava.so` and chain-loading `libjvm.so` MaxineVM fails (unless
`libjava.so` sets `DT_RUNPATH`, which it should NOT anymore).  To
overcome this issue we set `LD_LIBRARY_PATH` accordingly. Consequently
the use of `-rpath` when linking maxvm (see
`com.oracle.max.vm.native/platform/platform.mk`) is redundant now.