[test][fix][refactor] Tested and fixed test_movImmediate()
test_movImmediate() was failing because movImmediate doesn't properly sign-extend 16-bit integers.
Deprecated movImmediate() from the assembler and refactored it in the macro-assembler class as a mov() wrapper (which knows how to handle sign extension).
Needs to be further tested to ensure that the functionality of movImmediate() is covered by mov().