[fix][test][armv7] Patch trampoline call sites to avoid inf. loops
Fixup static trampoline's direct call to
patchStaticTrampolineCallSiteARMV7 to avoid infinite loops.  Note that
the patchStaticTrampolineCallSiteARMV7 is not copied to the codeBuffer,
because of its big chain of dependencies on other classes and methods
that would need to be copied as well.  That said the actual patching is
never performed, braking tests that rely on it.