[feat][aarch64][test] Generate code to fix SP in Aarch64CodeWriter
As arguments get pushed on the stack before calling the generated code
we need a way to restore the SP.  Placing such code inside the code
array won't work because the code array contains a return
instruction (generated by the JIT compiler for compiled method) and
returns before reaching any code we place at the end.  This patch fixes
this issue by restoring the SP in c_entry after the compiled method
returns.