[fix][aarch64] Prevent code execution during patching
There are cases where while patching a call site the code gets executed
by a different core.  Since a callsite consists of multiple instructions
there is no way to make the patching atomic.  As a result, we introduce
a method that uses 2 trampolines.  Any given time, only one of the two
trampolines can be in use, thus we can safely patch the other.  After
patching the trampoline we atomically modify the branch immediate that
directs the code execution to one of the two trampolines.