[fix][jsr292] Fix invokehandle for not resolvable w/o classloading cases
There are cases (see test.jsr292.MethodHandles_interface02) that fail to
resolve a classMethodref without class loading. In such cases we used to
fall back to the templates, which however assume only virtualmethods and
fail to catch methodhandle cases. This patch unconditionally tries to
resolve the method references to avoid the above issue.