[docker] Instead of mounting directories use rsync
Mounting host directories to containers and compiling from the container
appears to be very slow on MacOS (on Windows the overhead is less but
still noteable).  Using rsync to transfer only changes to the container
and then using the files "natively" in the container is much faster and
can be automated using fswatch.

Additionally not having to mount directories, we no longer need to use
the uid and gid of the host user, allowing us to use the root user as is
standard in docker.