[feat][aarch64] Support hybrid callsites
Hybrid call-sites allow a single branch call site with an out-of-line
trampoline. If a target can be reached from a call-site with a single
branch then that is used. Otherwise the trampoline is patched and the
call-site branch is directed to the trampoline.

Raise an exception if AbstractAssembler#trampolines() is called for
architectures that do not employ trampolines.