[fix] Ensure instruction stream synchronisation
Simplify the calculation of the extent of the memory segment containing
a compiled method and include any trampoline array in that calculation.
The call to cleanCache is moved from the initial code copying. It is the
responsibility of the concrete sub classes to call cleanCache after
any call-sites are fixed up. This reduces the number of cache maintenance
events and ensures stability.