[fix] Fix NUMA profiling
1. Use ETLA only to store the PROFILER_TLA thread local. This ensures
that profiling will happen only while threads are in normal execution
mode (safepoints enabled)

2. Since we skip the vmOperationThread when enabling the profiling we
don't need to pause and restart profiling when GCing. Note that this is
based on the assumption that GC is handles by the vmOperationThread and
that any GC work performed by application threads is happening when
safepoints are not enabled.

3. When enabling profiling because of the GC threshold, at the
enabling of profiling all application threads are frozen so we can use a
simple store, since there can't be data races.
On the contrary, when enabling/disabling profiling based on a
FlareObject, then the enabling happens from a single thread while the
rest of the threads are active. In that case we need a mechanism to
guarantee that the changes won't be overwritten, thus we use compare and
swap to manage the profiling state both in the thread enabling the
profiling and in the threads that check the profiling state.

4. Since we cannot control nor monitor the Finalizer thread, in the case
of -XX:+NUMAProfilerIncludeFinalization we don't do any deallocation of
the profiler resources to ensure that the Finalizer thread can continue
to profile accesses till the very end.