[HACK][fix] Prevent NUMAProfiler being invoked in T1X templates
NUMAProfiler needs to be implemented at the T1X level and not inside its
templates that are compiled by C1X. The main reason is that NUMAProfiler
is invoked using `callThroughStub` which is not safe to transfer
execution from T1X compiled code to C1X compiled code (i.e. the
NUMAProfiler)

This approach is not optimal since it touches the RiXirGenerator making
it "template-aware". Ideally the XIR should remain oblivious regarding
tiered compilation.